Add `TreeViewHierarchicalSelectBehavior`
